requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.12.3
FlagEmbedding==1.3.2
jieba==0.42.1
numpy==1.26.4
openai==1.54.4
pymilvus==2.5.0
PyYAML==6.0.2
readability-lxml==0.8.1
requests==2.32.3
rouge==1.0.1
scikit-learn==1.5.2
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
